+++ src/js/controllers/contactController.js
@@ -30,68 +30,73 @@
  * 
  */
 
+// import enableButton from '../utils/enableButton'
+// import { emailFeedback, nameFeedback } from '../utils/feedback'
+import {validName, validEmail, validComment} from '../utils/validFields'
+// import updateClase from '../utils/updateClass'
+
 // Funciones
 
 // Inicio
 function contactController() {
 
     // Actualizar la clase (valid - invalid)
-    let updateClase = (node, condition) => {
+    // let updateClase = (node, condition) => {
 
-        if (condition) {
-            node.addClass('is-valid')
-            node.removeClass('is-invalid')
-            enableButton()
-        } else {
-            node.addClass('is-invalid')
-            node.removeClass('is-valid')
-        }
-        enableButton()
-    }
+    //     if (condition) {
+    //         node.addClass('is-valid')
+    //         node.removeClass('is-invalid')
+    //         enableButton()
+    //     } else {
+    //         node.addClass('is-invalid')
+    //         node.removeClass('is-valid')
+    //     }
+    //     enableButton()
+    // }
 
     // Otorgar mensaje de feedback por nombre
-    let nameFeedback = (node, condition) => {
+    // let nameFeedback = (node, condition) => {
 
-        let feedbackNode = node.next()
+    //     let feedbackNode = node.next()
 
-        if (condition) {
-            feedbackNode.text('Nombre válido').addClass('text-success').removeClass('text-danger')
-        } else {
-            feedbackNode.text('Por favor, ingrese un nombre válido').addClass('text-danger').removeClass('text-success')
-        }
+    //     if (condition) {
+    //         feedbackNode.text('Nombre válido').addClass('text-success').removeClass('text-danger')
+    //     } else {
+    //         feedbackNode.text('Por favor, ingrese un nombre válido').addClass('text-danger').removeClass('text-success')
+    //     }
 
-    }
+    // }
 
     // Otorgar mensaje de feedback por email
-    let emailFeedback = (node, condition) => {
+    // let emailFeedback = (node, condition) => {
 
-        let feedbackNode = node.next()
+    //     let feedbackNode = node.next()
 
-        let hasNotHash = node.val().indexOf('@') == -1
-        let hasNotDotCom = node.val().indexOf('.com') == -1
-        let hasNone = hasNotDotCom && hasNotHash
+    //     let hasNotHash = node.val().indexOf('@') == -1
+    //     let hasNotDotCom = node.val().indexOf('.com') == -1
+    //     let hasNone = hasNotDotCom && hasNotHash
 
-        if (condition) {
-            feedbackNode.text('Email válido')
-                .addClass('text-success')
-                .removeClass('text-danger')
+    //     if (condition) {
+    //         feedbackNode.text('Email válido')
+    //             .addClass('text-success')
+    //             .removeClass('text-danger')
 
-        } else if (hasNone) {
-            feedbackNode.text('Por favor, ingrese un email válido (@ .com)')
-                .addClass('text-danger')
-                .removeClass('text-success')
+    //     } else if (hasNone) {
+    //         feedbackNode.text('Por favor, ingrese un email válido (@ .com)')
+    //             .addClass('text-danger')
+    //             .removeClass('text-success')
 
-        } else if (hasNotDotCom) {
-            feedbackNode.text('Por favor, ingrese un email válido (.com)')
-                .addClass('text-danger')
-                .removeClass('text-success')
+    //     } else if (hasNotDotCom) {
+    //         feedbackNode.text('Por favor, ingrese un email válido (.com)')
+    //             .addClass('text-danger')
+    //             .removeClass('text-success')
 
-        } else if (hasNotHash) {
-            feedbackNode.text('Por favor, ingrese un email válido (@)')
-                .addClass('text-danger')
-                .removeClass('text-success')
-        }
-    }
+    //     } else if (hasNotHash) {
+    //         feedbackNode.text('Por favor, ingrese un email válido (@)')
+    //             .addClass('text-danger')
+    //             .removeClass('text-success')
+    //     }
+    // }
 
     // Ejecutar Nombre
     $('#inputName').one('blur', function () {
@@ -104,14 +109,14 @@ function contactController() {
     })
 
     // Validar Nombre
-    let validName = (_e) => {
+    // let validName = (_e) => {
 
-        let validation = typeof _e.val() === 'string'
-            && _e.val().length > 1
+    //     let validation = typeof _e.val() === 'string'
+    //         && _e.val().length > 1
 
-        updateClase(_e, validation)
-        nameFeedback(_e, validation)
-    }
+    //     updateClase(_e, validation)
+    //     nameFeedback(_e, validation)
+    // }
 
     // Ejecutar Email
     $('#inputEmail').one('blur', function () {
@@ -121,14 +126,14 @@ function contactController() {
     })
 
     // Validar Email
-    let validEmail = (_e) => {
+    // let validEmail = (_e) => {
 
-        let validation = _e.val().indexOf('@') > -1
-            && _e.val().indexOf('.com') > -1
+    //     let validation = _e.val().indexOf('@') > -1
+    //         && _e.val().indexOf('.com') > -1
 
-        updateClase(_e, validation)
-        emailFeedback(_e, validation)
-    }
+    //     updateClase(_e, validation)
+    //     emailFeedback(_e, validation)
+    // }
 
     // Ejecutar Comentario
     $('#inputComment').one('blur', function () {
@@ -141,34 +146,34 @@ function contactController() {
     })
 
     // Validar Comentario
-    let validComment = (_e) => {
+    // let validComment = (_e) => {
 
-        let validation = _e.val() != null
-            && _e.val().length > 0
-            && _e.val() != ''
+    //     let validation = _e.val() != null
+    //         && _e.val().length > 0
+    //         && _e.val() != ''
 
-        updateClase(_e, validation)
-    }
+    //     updateClase(_e, validation)
+    // }
 
     // Habilitar el Botón Enviar
-    let enableButton = () => {
+    // let enableButton = () => {
 
-        let button = $('#button')
+    //     let button = $('#button')
 
-        let name = $('#inputName').hasClass('is-valid')
-        let email = $('#inputEmail').hasClass('is-valid')
-        let comment = $('#inputComment').hasClass('is-valid')
+    //     let name = $('#inputName').hasClass('is-valid')
+    //     let email = $('#inputEmail').hasClass('is-valid')
+    //     let comment = $('#inputComment').hasClass('is-valid')
 
-        let enableIt = name && email && comment
+    //     let enableIt = name && email && comment
 
-        if (enableIt) {
-            console.log('entro')
-            button.attr("disabled", false)
-            console.log(button)
-        } else {
-            button.attr("disabled", true)
-        }
-    }
+    //     if (enableIt) {
+    //         console.log('entro')
+    //         button.attr("disabled", false)
+    //         console.log(button)
+    //     } else {
+    //         button.attr("disabled", true)
+    //     }
+    // }
 }
 
 export default contactController
