+++ src/js/utils/updateClass.js
@@ -1,0 +1,16 @@
+import enableButton from '../utils/enableButton'
+
+let updateClase = (node, condition) => {
+
+    if (condition) {
+        node.addClass('is-valid')
+        node.removeClass('is-invalid')
+        enableButton()
+    } else {
+        node.addClass('is-invalid')
+        node.removeClass('is-valid')
+    }
+    enableButton()
+}
+
+export default updateClase
